- added @SuppressWarnings to unused servlet method parameters - removed unnecessary casts - removed unnecessary throw statements

+++ source/net/yacy/document/importer/OAIPMHImporter.java
@@ -115,7 +115,7 @@ public class OAIPMHImporter extends Thread implements Importer, Comparable<OAIPM
     }
 
     public int speed() {
-        return (int) (1000L * ((long) count()) / runningTime());
+        return (int) (1000L * (count()) / runningTime());
     }
     
     @Override
